+++ OMC/Infrastructure/WebApi/EventsHandler/Services/Responding/v2/NotifyCallbackResponder.cs
@@ -53,6 +53,14 @@ namespace EventsHandler.Services.Responding.v2
                 DeliveryReceipt callback = Serializer.Deserialize<DeliveryReceipt>(json);
                 FeedbackTypes status = callback.Status.ConvertToFeedbackStatus();
 
+                if (string.IsNullOrEmpty(callback.Reference))
+                {
+                    return new AcceptedResult
+                    {
+                        Value = new { message = "No reference found. Unable to send telemetry." }
+                    };
+                }
+
                 LogContactRegistration(callback, status);
 
                 HttpRequestResponse? informResult = null;
@@ -73,7 +81,7 @@ namespace EventsHandler.Services.Responding.v2
                     {
                         StatusCode = informResult.Value.IsFailure
                             ? StatusCodes.Status500InternalServerError
-                            : StatusCodes.Status200OK
+                            : StatusCodes.Status201Created
                     };
                 }
 
@@ -84,7 +92,7 @@ namespace EventsHandler.Services.Responding.v2
                     message = "Callback processed successfully (no telemetry sent)."
                 })
                 {
-                    StatusCode = StatusCodes.Status200OK
+                    StatusCode = StatusCodes.Status202Accepted
                 };
             }
             catch (Exception ex)
@@ -166,6 +174,7 @@ namespace EventsHandler.Services.Responding.v2
                         {
                             NotifyMethods.Email => configuration.AppSettings.Variables.UxMessages.Email_Success_Subject(),
                             NotifyMethods.Sms => configuration.AppSettings.Variables.UxMessages.SMS_Success_Subject(),
+                            NotifyMethods.Letter => configuration.AppSettings.Variables.UxMessages.Letter_Success_Body(),
                             _ => string.Empty
                         }
                         : subject  // Use the subject from the original notification if available
@@ -176,6 +185,7 @@ namespace EventsHandler.Services.Responding.v2
                     {
                         NotifyMethods.Email => configuration.AppSettings.Variables.UxMessages.Email_Failure_Subject(),
                         NotifyMethods.Sms => configuration.AppSettings.Variables.UxMessages.SMS_Failure_Subject(),
+                        NotifyMethods.Letter => configuration.AppSettings.Variables.UxMessages.Letter_Failure_Body(),
                         _ => string.Empty
                     },